more sundile_pack updates

+++ tools/src/bin/sundile_pack.rs
@@ -52,11 +52,9 @@ struct Args {
     #[arg(long, short, default_value = "./")]
     target_directory: String,
     /// Directory to output the built package.
-    #[arg(long, short = 'o', default_value = "./server/pkg")]
+    /// This is relative to the target directory!
+    #[arg(long, short = 'o', default_value = "server/pkg")]
     out_dir: String,
-    /// Name of the output module.
-    #[arg(long, short = 'n', default_value = "target")]
-    out_name: String,
 
     /// Create a development build. Enable debug info, and disable optimizations.
     #[arg(long)]
@@ -67,6 +65,10 @@ struct Args {
     /// Create a profiling build. Enable optimizations and debug info.
     #[arg(long)]
     profiling: bool,
+
+    /// Create a .zip archive for easy upload to sites like itch.io.
+    #[arg(long)]
+    zip: bool,
 
     // /// Arguments to be passed to wasm-pack.
     // #[arg(long,short)]
@@ -88,12 +90,13 @@ enum BuildError {
     Custom(String),
 }
 
+// Calls wasm-pack and adds the included index.html.
 fn build<'a>(args: &Args) -> Result<(), BuildError> {
     let mut cmd = std::process::Command::new("wasm-pack");
     cmd.arg("build")
         .arg("--target=web")
         .arg(format!("--out-dir={}", args.out_dir))
-        .arg(format!("--out-name={}", args.out_name));
+        .arg("--out-name=target");
 
     if args.dev {
         cmd.arg("--dev");
@@ -104,14 +107,21 @@ fn build<'a>(args: &Args) -> Result<(), BuildError> {
     if args.profiling {
         cmd.arg("--profiling");
     }
+    if !(args.dev || args.release || args.profiling) {
+        println!("WARNING: wasm-bindgen compiles release builds by default.");
+        println!("To enable development builds, pass '--dev'");
+    }
 
     cmd.arg(args.target_directory.clone())
         .args(args.cargo_args.clone());
 
     match cmd.spawn()?.wait_with_output()?.status.success() {
         true => {
-            File::create(format!("{}/index.html", args.out_dir))?
-                .write(include_bytes!("index.html"))?;
+            File::create(format!(
+                "{}/{}/index.html",
+                args.target_directory, args.out_dir
+            ))?
+            .write(include_bytes!("index.html"))?;
 
             Ok(())
         }
@@ -119,9 +129,10 @@ fn build<'a>(args: &Args) -> Result<(), BuildError> {
     }
 }
 
-fn write_zip<W: Write + Seek>(dir: String, zip: &mut ZipWriter<W>) -> Result<(), BuildError> {
-    let pkg_dir = read_dir(&dir)?;
-    zip.add_directory(&dir, Default::default())?;
+// Recursively zips the package directory for easy upload to e.g. itch.io
+fn write_zip<W: Write + Seek>(dir: &String, zip: &mut ZipWriter<W>) -> Result<(), BuildError> {
+    let pkg_dir = read_dir(dir)?;
+    zip.add_directory(dir, Default::default())?;
 
     for file in pkg_dir {
         let entry = file?;
@@ -135,7 +146,7 @@ fn write_zip<W: Write + Seek>(dir: String, zip: &mut ZipWriter<W>) -> Result<(),
                     path
                 )))?;
             let new_dir = dir.clone() + "/" + file_name;
-            write_zip(new_dir, zip)?;
+            write_zip(&new_dir, zip)?;
         } else {
             let file_name = entry.file_name().into_string().map_err(|e| {
                 BuildError::Custom(format!("Unable to convert path string {:?} to UTF-8", e))
@@ -154,24 +165,28 @@ fn main() -> Result<(), BuildError> {
 
     build(&args)?;
 
-    let archive_path = format!("{}/itch-pack.zip", args.out_dir);
-    let archive = File::create(&archive_path)?;
-    let mut zip = ZipWriter::new(archive);
-    match write_zip("./server".into(), &mut zip) {
-        Err(e) => {
-            // Finish and remove file. Don't want a half-packaged thing.
-            zip.finish()?;
-            remove_file(&archive_path)?;
-            Err(e)
+    let archive_path = format!("{}/{}/itch-pack.zip", args.target_directory, args.out_dir);
+    if args.zip {
+        let archive = File::create(&archive_path)?;
+        let mut zip = ZipWriter::new(archive);
+        match write_zip(&archive_path, &mut zip) {
+            Ok(_) => {
+                zip.finish()?;
+                println!("Archive successfully created!");
+            }
+            Err(e) => {
+                // Finish and remove file. Don't want a half-packaged thing.
+                zip.finish()?;
+                remove_file(&archive_path)?;
+                return Err(e);
+            }
         }
-        _ => {
-            zip.finish()?;
-            println!("Archive successfully created!");
-            println!(
-                "(Tip: Run `npx http-server {} -o` to test locally.)",
-                args.out_dir
-            );
-            Ok(())
-        }
-    }
+    };
+
+    println!(
+        "(Tip: Run `npx http-server {}/{} -o index.html` to test locally.)",
+        args.target_directory, args.out_dir,
+    );
+
+    Ok(())
 }
